Add some fields to nodeinfo

+++ src/baldr/graphid.cc
@@ -4,68 +4,74 @@
 #include "baldr/graphid.h"
 
 namespace {
-
-constexpr unsigned int kMaxGraphTileId = 16777215;
+// Maximum number of tiles supported.
+constexpr uint32_t kMaxGraphTileId = 16777215;
 
 // Maximum of 8 (0-7) graph hierarchies are supported.
-constexpr unsigned int kMaxGraphHierarchy = 7;
+constexpr uint32_t kMaxGraphHierarchy = 7;
 
 // Maximum unique identifier within a graph hierarchy.
 constexpr uint64_t kMaxGraphId = 68719476735;
-
 }
 
 namespace valhalla {
 namespace baldr {
 
+// Default constructor
 GraphId::GraphId() {
   graphid_.v = 0;
 }
 
-GraphId::GraphId(const unsigned int tileid, const unsigned int level,
-                 const unsigned int id) {
+// Constructor with values for each field of the GraphId.
+GraphId::GraphId(const uint32_t tileid, const uint32_t level,
+                 const uint64_t id) {
   Set(tileid, level, id);
 }
 
+// Copy constructor
 GraphId::GraphId(const GraphId& g) {
   Set(g.tileid(), g.level(), g.id());
 }
 
+// Get a 64 bit value of the composite GraphId
 uint64_t GraphId::value() const {
   return graphid_.v;
 }
 
-unsigned int GraphId::tileid() const {
+// Get the tile Id
+uint32_t GraphId::tileid() const {
   return graphid_.fields.tileid;
 }
 
-unsigned int GraphId::level() const {
+// Get the hierarchy level
+uint32_t GraphId::level() const {
   return graphid_.fields.level;
 }
 
-unsigned int GraphId::id() const {
+// Get the Id
+uint64_t GraphId::id() const {
   return graphid_.fields.id;
 }
 
-void GraphId::Set(const unsigned int tileid, const unsigned int level,
-                  const unsigned int id) {
+// Set the fields of the GraphId
+void GraphId::Set(const uint32_t tileid, const uint32_t level,
+                  const uint64_t id) {
   graphid_.fields.tileid = (tileid < kMaxGraphTileId) ? tileid : 0;
   graphid_.fields.level = (level < kMaxGraphHierarchy) ? level : 0;
   graphid_.fields.id = (id < kMaxGraphId) ? id : 0;
 }
 
-/**
- * Post increments the id.
- */
+// Post increments the id.
 void GraphId::operator ++(int) {
   graphid_.fields.id++;
 }
 
-// TODO - could this be simplified using unions in the struct?
+// Comparison for sorting
 bool GraphId::operator <(const GraphId& rhs) const {
   return graphid_.v < rhs.graphid_.v;
 }
 
+// Equality operator
 bool GraphId::operator ==(const GraphId& rhs) const {
   return graphid_.v == rhs.graphid_.v;
 }
@@ -78,6 +84,7 @@ std::size_t GraphId::HashCode() const {
 
   return seed;
 }
+
 
 }
 }
